Fix the missing regex_replace function issue
