+++ intel_extension_for_transformers/llm/runtime/graph/application/common.h
@@ -23,6 +23,7 @@
 #include <tuple>
 #include <vector>
 #include <random>
+#include <regex>
 #include <thread>
 
 #include "core/data_types.h"
@@ -189,3 +190,9 @@ void console_init(console_state& con_st);
 void console_cleanup(console_state& con_st);
 void console_set_color(console_state& con_st, console_color_t color);
 bool console_readline(console_state& con_st, std::string& line);
+
+std::string build_prompt_glm2(const std::vector<std::string>& history);
+std::string build_prompt_glm1(const std::vector<std::string>& history);
+static std::string regex_replace(const std::string &input, const std::regex &regex,
+                                 std::function<std::string(const std::smatch &)> format);
+std::string postprocess(const std::string &text);
